refactor: extract isPrivateUrl to shared utility
